+++ docs-web-common/src/main/java/com/sismics/rest/util/RestUtil.java
@@ -1,9 +1,5 @@
 package com.sismics.rest.util;
 
-import com.sismics.docs.core.constant.AclType;
-import com.sismics.docs.core.constant.PermType;
-import com.sismics.docs.core.dao.AclDao;
-import com.sismics.docs.core.dao.dto.AclDto;
 import com.sismics.docs.core.model.jpa.File;
 import com.sismics.docs.core.util.DirectoryUtil;
 import com.sismics.docs.core.util.FileUtil;
@@ -11,11 +7,9 @@ import com.sismics.rest.exception.ServerException;
 import com.sismics.util.JsonUtil;
 
 import javax.json.Json;
-import javax.json.JsonArrayBuilder;
 import javax.json.JsonObjectBuilder;
 import java.io.IOException;
 import java.nio.file.Files;
-import java.util.List;
 
 /**
  * Rest utilities.
@@ -23,28 +17,6 @@ import java.util.List;
  * @author bgamard
  */
 public class RestUtil {
-    /**
-     * Add ACLs to a JSON response.
-     *
-     * @param json JSON
-     * @param sourceId Source ID
-     * @param targetIdList List of target ID
-     */
-    public static void addAcls(JsonObjectBuilder json, String sourceId, List<String> targetIdList) {
-        AclDao aclDao = new AclDao();
-        List<AclDto> aclDtoList = aclDao.getBySourceId(sourceId, AclType.USER);
-        JsonArrayBuilder aclList = Json.createArrayBuilder();
-        for (AclDto aclDto : aclDtoList) {
-            aclList.add(Json.createObjectBuilder()
-                    .add("perm", aclDto.getPerm().name())
-                    .add("id", aclDto.getTargetId())
-                    .add("name", JsonUtil.nullable(aclDto.getTargetName()))
-                    .add("type", aclDto.getTargetType()));
-        }
-        json.add("acls", aclList)
-                .add("writable", aclDao.checkPermission(sourceId, PermType.WRITE, targetIdList));
-    }
-
     /**
      * Transform a File into its JSON representation
      * @param fileDb a file
